Refactor and cleanup utils, add protocol `DecodingInputsType` for DecoderInputs
* Refactor and cleanup Utils

* Fix @available

* Fix warnings in Example app

* Make MLMultiArray extension public

* Refactor transcription utils under TranscriptionUtilities

* Add docs to utility functions

* Introduce AudioProcessorOutputType for polymorphism

* Add voiceActivityAsync

* Add DecodingInputsType for arbitrary types

* Extra decoding options

* Move prefill masks to protocol

* Dynamic prefill cache reset size

---------
