+++ Sources/WhisperKit/Utilities/ResultWriter.swift
