+++ Sources/WhisperKitCLI/TranscribeCLI.swift
@@ -54,7 +54,7 @@ struct TranscribeCLI: AsyncParsableCommand {
     }
 
     private func transcribe() async throws {
-        let resolvedAudioPaths = cliArguments.audioPath.map { resolveAbsolutePath($0) }
+        let resolvedAudioPaths = cliArguments.audioPath.map { FileManager.resolveAbsolutePath($0) }
         for resolvedAudioPath in resolvedAudioPaths {
             guard FileManager.default.fileExists(atPath: resolvedAudioPath) else {
                 throw CocoaError.error(.fileNoSuchFile)
@@ -101,14 +101,14 @@ struct TranscribeCLI: AsyncParsableCommand {
 
         // Log timings for full transcription run
         if cliArguments.verbose {
-            let mergedResult = mergeTranscriptionResults(allSuccessfulResults)
+            let mergedResult = TranscriptionUtilities.mergeTranscriptionResults(allSuccessfulResults)
             mergedResult.logTimings()
         }
 
         for (audioPath, result) in zip(resolvedAudioPaths, transcribeResult) {
             do {
                 let partialResult = try result.get()
-                let mergedPartialResult = mergeTranscriptionResults(partialResult)
+                let mergedPartialResult = TranscriptionUtilities.mergeTranscriptionResults(partialResult)
                 processTranscriptionResult(audioPath: audioPath, transcribeResult: mergedPartialResult)
             } catch {
                 print("Error when transcribing \(audioPath): \(error)")
@@ -166,7 +166,7 @@ struct TranscribeCLI: AsyncParsableCommand {
         guard let audioPath = cliArguments.audioPath.first else {
             throw CocoaError.error(.fileNoSuchFile)
         }
-        let resolvedAudioPath = resolveAbsolutePath(audioPath)
+        let resolvedAudioPath = FileManager.resolveAbsolutePath(audioPath)
         guard FileManager.default.fileExists(atPath: resolvedAudioPath) else {
             throw CocoaError.error(.fileNoSuchFile)
         }
@@ -215,7 +215,7 @@ struct TranscribeCLI: AsyncParsableCommand {
 
                     if let prevResult = prevResult {
                         prevWords = prevResult.allWords.filter { $0.start >= lastAgreedSeconds }
-                        let commonPrefix = findLongestCommonPrefix(prevWords, hypothesisWords)
+                        let commonPrefix = TranscriptionUtilities.findLongestCommonPrefix(prevWords, hypothesisWords)
                         if cliArguments.verbose {
                             print("[whisperkit-cli] Prev \"\((prevWords.map { $0.word }).joined())\"")
                             print("[whisperkit-cli] Next \"\((hypothesisWords.map { $0.word }).joined())\"")
@@ -257,10 +257,10 @@ struct TranscribeCLI: AsyncParsableCommand {
             }
         }
 
-        let final = lastAgreedWords + findLongestDifferentSuffix(prevWords, hypothesisWords)
+        let final = lastAgreedWords + TranscriptionUtilities.findLongestDifferentSuffix(prevWords, hypothesisWords)
         confirmedWords.append(contentsOf: final)
 
-        let mergedResult = mergeTranscriptionResults(results, confirmedWords: confirmedWords)
+        let mergedResult = TranscriptionUtilities.mergeTranscriptionResults(results, confirmedWords: confirmedWords)
 
         processTranscriptionResult(audioPath: audioPath, transcribeResult: mergedResult)
     }
